Fix a bug in determining the output path.
I'm a little worried that I'm breaking something, because I don't
understand why I wrote the code the way it was.  It seems like I was
using str() to guard against being given a non-string path, but I did it
in a way that made the default path name "None", which is clearly wrong.
What I don't understand is why this function would ever be given a
non-string path, or why it wouldn't be the responsibility of the caller
to do the type-casting if they did want to use a non-string.

Anyways, I guess I'll eventually notice if anything breaks. :-)